draw me:
- drawing first element and add it to main.js
